+++ app/src/main/java/chpc/UI/controllers/AddPanelCommand.java
@@ -1,4 +1,6 @@
-package chpc.UI;
+package chpc.UI.controllers;
+
+import chpc.UI.views.MainUI;
 
 import javax.swing.JComponent;
