+++ app/src/main/java/chpc/UI/views/VisualizationSelectionWindow.java
@@ -1,4 +1,4 @@
-package chpc.UI;
+package chpc.UI.views;
 
 import java.util.Set;
