+++ app/src/main/java/chpc/visualizations/views/Line.java
@@ -1,4 +1,4 @@
-package chpc.visualizations;
+package chpc.visualizations.views;
 
 import org.jfree.chart.ChartFactory;
 import org.jfree.chart.JFreeChart;
